+++ types/storybook__react/tsconfig.json
@@ -19,7 +19,7 @@
                 "storybook__react"
             ],
             "@storybook/react/demo": [
-                "storybook__react"
+                "storybook__react/demo"
             ]
         },
         "types": [],
@@ -28,6 +28,7 @@
     },
     "files": [
         "index.d.ts",
+        "demo/index.d.ts",
         "storybook__react-tests.tsx"
     ]
 }
